+++ components/Loader/index.js
@@ -1,0 +1,36 @@
+import styled from "styled-components";
+
+const Loader = styled.div`
+  display: block;
+  width: 130px;
+  height: 4px;
+  border-radius: 30px;
+  background-color: rgba(0, 0, 0, 0.2);
+  position: relative;
+
+  :before {
+    content: "";
+    position: absolute;
+    background: var(--primary-color);
+    top: 0;
+    left: 0;
+    width: 0%;
+    height: 100%;
+    border-radius: 30px;
+    animation: moving 1s ease-in-out infinite;
+  }
+
+  @keyframes moving {
+    50% {
+      width: 100%;
+    }
+
+    100% {
+      width: 0;
+      right: 0;
+      left: unset;
+    }
+  }
+`;
+
+export default Loader;
