+++ components/StyledList/index.js
@@ -1,0 +1,12 @@
+import styled from "styled-components";
+
+const StyledList = styled.ul`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 90%;
+  //height: 254px;
+  padding: 0;
+`;
+
+export default StyledList;
